vendor_backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
Pillow==10.4.0
django-channels==4.1.0
channels-redis==4.2.0
python-decouple==3.8